+++ sdk/core/Azure.Core/src/Pipeline/DefaultRetryPolicy.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Azure.Core.Pipeline;
-
-namespace Azure.Core
+namespace Azure.Core.Pipeline
 {
     internal class DefaultRetryPolicy : RetryPolicy
     {
